+++ src/pages/Home/UserMainPage.tsx
@@ -1,45 +1,15 @@
-import { useEffect, useState } from 'react';
-
-import { useNavigate } from 'react-router-dom';
 import 'slick-carousel/slick/slick-theme.css';
 import 'slick-carousel/slick/slick.css';
 
 import Banner from '@/components/Banner';
-import Button from '@/components/Button';
-import ROUTES from '@/constants/routes';
-import MatchingStatusCard from '@/features/Category/components/MatchingStatusCard';
-import { dummyMatchings } from '@/features/Category/data/dummy';
+import RealtimeMatchingStatus from '@/components/RealtimeMatchingStatus';
 import RequestSlider from '@/features/home/components/RequestSlider';
 import UserSearchSection from '@/features/home/components/UserSearchSection';
 import mockRequests from '@/features/home/data/dummy';
 import { useUserRoleStore } from '@/store/useUserRoleStore';
 
 const UserMainPage = () => {
-  const { isLoggedIn, isExpert } = useUserRoleStore();
-  const [cardCount, setCardCount] = useState(8);
-  const navigate = useNavigate();
-
-  const handleRequestWriteClick = () => {
-    if (!isLoggedIn) {
-      alert('로그인이 필요합니다.');
-      navigate(ROUTES.AUTH.LOGIN);
-    } else if (isExpert) {
-      alert('전문가 계정은 요청서를 작성할 수 없습니다.');
-    } else navigate(ROUTES.REQUESTS.NEW);
-  };
-
-  // 화면 폭에 따라 카드 개수 조정
-  useEffect(() => {
-    const update = () => {
-      const w = window.innerWidth;
-      if (w >= 1536) setCardCount(8);
-      else if (w >= 1280) setCardCount(6);
-      else setCardCount(4);
-    };
-    update();
-    window.addEventListener('resize', update);
-    return () => window.removeEventListener('resize', update);
-  }, []);
+  const { isLoggedIn } = useUserRoleStore();
 
   return (
     <div className="flex flex-col items-center justify-center">
@@ -51,31 +21,10 @@ const UserMainPage = () => {
         </div>
       )}
 
-      <section className="mt-[200px] mb-[200px] flex max-w-[1480px] flex-col gap-9 px-4 sm:w-[720px] lg:w-[720px] xl:w-[1080px] 2xl:w-[1480px]">
-        <h2 className="font-[Pretendard Variable] leading-[40px] font-extrabold text-black not-italic sm:text-[24px] xl:text-[30px]">
-          실시간 매칭 현황
-        </h2>
+      <div className="my-[200px]">
+        <RealtimeMatchingStatus />
+      </div>
 
-        <div className="mt-9 grid grid-cols-1 gap-x-[20px] gap-y-[16px] md:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4">
-          {dummyMatchings.slice(0, cardCount).map((match, idx) => (
-            <MatchingStatusCard key={idx} {...match} />
-          ))}
-        </div>
-
-        {/* 오른쪽 아래 정렬 */}
-        <div className="flex justify-end">
-          <Button
-            label="요청서 작성"
-            width="w-[189px]"
-            height="h-[42px]"
-            text="text-[15px] font-semibold text-white"
-            onClick={handleRequestWriteClick}
-            className="cursor-pointer"
-          >
-            요청서 작성
-          </Button>
-        </div>
-      </section>
       <div className="mx-auto mb-[200px] max-w-[1480px] sm:w-[720px] lg:w-[720px] xl:w-[1080px] 2xl:w-[1480px]">
         <Banner />
       </div>
